+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# [6.1.0](https://github.com/invertase/react-native-firebase/compare/v6.0.4...v6.1.0) (2019-11-26)
+# Next
+
+## storage
+- added `snapshot` property to `Task` as per web spec.
 
 
 ### Bug Fixes